fix(playground): wrap nuqs client component in Suspense to fix prerender
